+++ sprint_2/src/components/Login/components_login/Generales/Contenido.js
