+++ sprint_2/src/components/Login/components_login/Generales/Logo.js
@@ -1,4 +1,4 @@
-import logo from "../assets-login/Logo_ITBANK.png"
+import logo from "../../assets-login/Logo_ITBANK.png"
 import { Link } from "react-router-dom"
 
 export function Logo(){
