+++ sprint_2/src/components/Login/components_login/Generales/Subtitulo.js
