+++ sprint_2/src/components/Login/components_login/Password/FormPassword.js
@@ -1,7 +1,7 @@
-import { Boton } from "./Boton"
-import { Input } from "./Input"
-import { Label } from "./Label"
-import { Subtitulo } from "./Subtitulo"
+import { Boton } from "../Generales/Boton"
+import { Input } from "../Generales/Input"
+import { Label } from "../Generales/Label"
+import { Subtitulo } from "../Generales/Subtitulo"
 
 export function FormPassword(){
     return (
